mcp_demo/requirements.txt
# MCP Server dependencies
fastmcp>=0.1.0
psycopg[binary]>=3.1.0
mysql-connector-python>=8.0.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
requests>=2.31.0

# For data types and validation
typing-extensions>=4.0.0

# For better error handling and logging
structlog>=23.0.0
